+++ src/assets/emotes.json
@@ -93,6 +93,5 @@
 	"check": {
 		"green": "<:greenTick:1005044096534069369>",
 		"red": "<:redTick:1005044095196086314>"
-	},
-	"rankTemplate": "https://cdn.discordapp.com/attachments/935802603957272576/1015190231428907068/rank.png"
+	}
 }
