+++ src/utils/bisect.ts
@@ -1,4 +1,4 @@
-export default function bisect<T>(array: T[], item: T): number {
+export function bisect<T>(array: T[], item: T): number {
 	let low = 0;
 	let mid;
 	let high = array.length;
